+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # RealmJoin Remediation Scripts Changelog
 
+## 2026-09-09
+- Added Enable Automatic Time Zone Service
+
 ## 2026-08-28
 - Added Prefer IPv4 over IPv6